requirements.txt
gitpython>=3.1.30
pyyaml>=6.0
requests>=2.28.0
pytest>=7.3.1
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
types-PyYAML>=6.0.12
types-requests>=2.28.11

